feat: move to own util file
